Refactor AuthScreen to use modular authentication components
