+++ src/components/auth/AuthForm.tsx
@@ -1,0 +1,90 @@
+import React, { useState } from 'react';
+import { View, Text, TextInput, StyleSheet, Pressable, ActivityIndicator } from 'react-native';
+import { colors } from '../../theme/colors';
+
+const AuthForm = ({ isLogin, signIn, signUp }: { isLogin: boolean, signIn: (email: string, password: string) => Promise<void>, signUp: (email: string, password: string) => Promise<void> }) => {
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState('');
+
+  const handleSubmit = async () => {
+    try {
+      setLoading(true);
+      setError('');
+      if (isLogin) {
+        await signIn(email.trim(), password);
+      } else {
+        await signUp(email.trim(), password);
+      }
+    } catch (e: any) {
+      setError(e.message);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  return (
+    <View>
+      <Text style={styles.title}>{isLogin ? '登入' : '註冊'}</Text>
+      <TextInput
+        style={styles.input}
+        placeholder="電子郵件"
+        value={email}
+        onChangeText={setEmail}
+        autoCapitalize="none"
+        keyboardType="email-address"
+      />
+      <TextInput
+        style={styles.input}
+        placeholder="密碼"
+        value={password}
+        onChangeText={setPassword}
+        secureTextEntry
+      />
+      {error ? <Text style={styles.error}>{error}</Text> : null}
+      <Pressable style={styles.submitButton} onPress={handleSubmit} disabled={loading}>
+        {loading ? (
+          <ActivityIndicator color="#fff" />
+        ) : (
+          <Text style={styles.submitText}>
+            {isLogin ? '登入' : '註冊'}
+          </Text>
+        )}
+      </Pressable>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  title: {
+    fontSize: 24,
+    fontWeight: 'bold',
+    marginBottom: 20,
+    color: colors.text,
+    textAlign: 'center',
+  },
+  input: {
+    backgroundColor: '#fff',
+    padding: 15,
+    borderRadius: 8,
+    marginBottom: 10,
+  },
+  error: {
+    color: 'red',
+    marginBottom: 10,
+  },
+  submitButton: {
+    backgroundColor: '#007AFF',
+    padding: 15,
+    borderRadius: 8,
+    marginBottom: 10,
+  },
+  submitText: {
+    color: '#fff',
+    textAlign: 'center',
+    fontWeight: 'bold',
+  },
+});
+
+export default AuthForm;
